+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.0
 
 * Rename `Graph` to `Gviz`.
+* Added `Graph` class which actually represents a DAG.
 * Ensure nodes are only added once.
 
 ## 0.0.1